feat: :art: code is modified
The code was modified by adding a new image using the "img" command. Aside from this, there were minor touches such as changing the color of a "column" and the color of some letters. In a brief summary, an image was added, a color was changed, and spelling errors were corrected.

GitHub link and an image of the "Campuslands" logo have been added.